Create Parser class for parsing user inputs.
