+++ src/main/java/Duke.java
@@ -7,12 +7,6 @@ import java.util.ArrayList;
 import java.util.Scanner;
 
 public class Duke {
-    /**
-     * Lists the list of commands.
-     */
-    public static void mismatch() {
-        Ui.reply("list of commands: list, mark, unmark, todo, deadline, event, delete");
-    }
 
     public static void main(String[] args) throws IOException {
         String logo = " ____        _        \n"
@@ -37,42 +31,6 @@ public class Duke {
             }
         }
         Ui.reply("What can I do for you?");
-        while (true) {
-            String line = Ui.readLine();
-            String[] arguments = Ui.parse(line);
-            if (arguments[0].equals("bye")) {
-                Ui.bye();
-                break;
-            }
-            if (arguments[0].equals("list")){
-                TaskList.list();
-                continue;
-            }
-            if (arguments[0].equals("mark")) {
-                TaskList.mark(arguments);
-                continue;
-            }
-            if (arguments[0].equals("unmark")) {
-                TaskList.unmark(arguments);
-                continue;
-            }
-            if (arguments[0].equals("todo")) {
-                TaskList.todo(arguments);
-                continue;
-            }
-            if (arguments[0].equals("deadline")) {
-                TaskList.deadline(arguments);
-                continue;
-            }
-            if (arguments[0].equals("event")) {
-                TaskList.event(arguments);
-                continue;
-            }
-            if (arguments[0].equals("delete")) {
-                TaskList.delete(arguments);
-                continue;
-            }
-            mismatch();
-        }
+        while (Parser.parseExecute(Ui.readLine()));
     }
 }
